python/juniper-mist-kentic-webhooks/pyproject.toml
[tool.poetry]
name = "juniper-mist-kentic-webhooks"
version = "0.1.0"
description = "Kentic webhook receiver to execute API call to Mist"
authors = ["Calvin Remsburg <cremsburg.dev@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.9.1"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
uvicorn = "^0.18.2"
uvloop = "^0.16.0"
httptools = "^0.4.0"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
flake8-black = "^0.3.3"
invoke = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
